Move deployment function to deployment files to work around a typing issue
CI was reverting while the function was located in the library files.
Moving it to the deployment files is a hack that fixes the issue.
I'm not really sure what's going on, but it looks like the type I need to use comes from type extensions and `declare module` so I'm not sure how to access them directly.

Relevant issue link:
https://github.com/wighawag/hardhat-deploy-ethers/issues/18
