+++ src/deploy/002_CowDaoAllocationModule.ts
@@ -1,6 +1,8 @@
 import { DeployFunction } from "hardhat-deploy/types";
 
-import { COW_DAO_SAFE, allocationModuleDeployFunction } from "../ts";
+import { COW_DAO_SAFE } from "../ts";
+
+import { allocationModuleDeployFunction } from "./001_AllocationModule";
 
 const deployCowDaoAuthenticator: DeployFunction =
   allocationModuleDeployFunction(COW_DAO_SAFE);
